+++ .ncurc.js
@@ -1,12 +1,9 @@
 module.exports = {
-  // upgrade: true,
+  upgrade: true,
   // target: 'newest',
   reject: [
-    'webpack-dev-server',
-    'eslint',
     'electron',
     'electron-builder',
     'chalk',
-    // 'eslint-config-standard'
   ]
 }