+++ build-config/renderer/webpack.config.dev.js
@@ -9,7 +9,6 @@ module.exports = merge(baseConfig, {
   mode: 'development',
   devtool: 'eval-source-map',
   plugins: [
-    new webpack.HotModuleReplacementPlugin(),
     new webpack.DefinePlugin({
       'process.env': {
         NODE_ENV: '"development"',