+++ src/main/rendererEvents/getEnvParams.js
@@ -1,6 +1,9 @@
-const { mainHandle, NAMES: { mainWindow: ipcMainWindowNames } } = require('../../common/ipc')
+const { mainHandle, mainOn, NAMES: { mainWindow: ipcMainWindowNames } } = require('../../common/ipc')
 
 mainHandle(ipcMainWindowNames.get_env_params, async(event, options) => {
-  return global.envParams.cmdParams
+  return global.envParams
 })
 
+mainOn(ipcMainWindowNames.clear_env_params_deeplink, () => {
+  global.envParams.deeplink = null
+})