+++ src/renderer/components/common/AudioVisualizer.vue
@@ -11,19 +11,19 @@ import { isPlay } from '@renderer/core/share/player'
 import { player as eventPlayerNames } from '@renderer/event/names'
 
 const themes = {
-  green: 'rgba(77,175,124,.1)',
-  blue: 'rgba(52,152,219,.1)',
-  yellow: 'rgba(233,212,96,.16)',
-  orange: 'rgba(245,171,53,.1)',
-  red: 'rgba(214,69,65,.08)',
-  pink: 'rgba(241,130,141,.1)',
-  purple: 'rgba(155,89,182,.1)',
-  grey: 'rgba(108,122,137,.1)',
-  ming: 'rgba(51,110,123,.1)',
-  blue2: 'rgba(79,98,208,.1)',
-  black: 'rgba(39,39,39,.26)',
-  mid_autumn: 'rgba(74,55,82,.05)',
-  naruto: 'rgba(87,144,167,.1)',
+  green: 'rgba(77,175,124,.16)',
+  blue: 'rgba(52,152,219,.16)',
+  yellow: 'rgba(233,212,96,.22)',
+  orange: 'rgba(245,171,53,.16)',
+  red: 'rgba(214,69,65,.12)',
+  pink: 'rgba(241,130,141,.16)',
+  purple: 'rgba(155,89,182,.14)',
+  grey: 'rgba(108,122,137,.16)',
+  ming: 'rgba(51,110,123,.14)',
+  blue2: 'rgba(79,98,208,.14)',
+  black: 'rgba(39,39,39,.4)',
+  mid_autumn: 'rgba(74,55,82,.1)',
+  naruto: 'rgba(87,144,167,.14)',
   happy_new_year: 'rgba(192,57,43,.1)',
 }
 export default {
@@ -42,6 +42,11 @@ export default {
     let isPlaying = false
     let animationFrameId
 
+    let num
+    let mult
+    const maxNum = 255
+    let frequencyAvg = 0
+
     const theme = useRefGetter('theme')
     // const setting = useRefGetter('setting')
     let themeColor = themes[theme.value || 'green']
@@ -49,7 +54,7 @@ export default {
       themeColor = themes[theme || 'green']
     })
 
-    // https://codepen.io/nfj525/pen/rVBaab
+    // https://developer.mozilla.org/zh-CN/docs/Web/API/AnalyserNode/smoothingTimeConstant
     const renderFrame = () => {
       animationFrameId = null
       if (isPlaying) animationFrameId = window.requestAnimationFrame(renderFrame)
@@ -60,8 +65,23 @@ export default {
 
       ctx.clearRect(0, 0, WIDTH, HEIGHT)
       // ctx.fillRect(0, 0, WIDTH, HEIGHT)
+      ctx.fillStyle = themeColor
 
       for (let i = 0; i < bufferLength; i++) {
+        mult = Math.floor(i / maxNum)
+        num = mult % 2 === 0 ? (i - maxNum * mult) : (maxNum - (i - maxNum * mult))
+        let spectrum = num > 90 ? 0 : dataArray[num + 20]
+        frequencyAvg += spectrum * 1.2
+      }
+      frequencyAvg /= bufferLength
+      frequencyAvg *= 1.4
+
+      frequencyAvg = frequencyAvg / maxNum
+      // ctx.scale(1, 1 + frequencyAvg)
+
+      for (let i = 0; i < bufferLength; i++) {
+        if (x > WIDTH) break
+
         barHeight = dataArray[i]
 
         // let r = barHeight + (25 * (i / bufferLength))
@@ -69,7 +89,7 @@ export default {
         // let b = 50
 
         // ctx.fillStyle = 'rgb(' + r + ',' + g + ',' + b + ')'
-        ctx.fillStyle = themeColor
+        barHeight = barHeight * frequencyAvg + barHeight * 0.42
         ctx.fillRect(x, HEIGHT - barHeight, barWidth, barHeight)
 
         x += barWidth